Bugfixes
Fix typo in README.md.
Due to the typo, I had set the environment variable as OPENAPI_API_KEY.
Using LLMVM_EXECUTOR="openai" LLMVM_MODEL="gpt-4o" was falling into
the else block and using an AnthropicExecutor since there was also an
ANTHROPIC_API_KEY in the environment. When "openai" is provided as an
executor, it should use an OpenAIExecutor or error out.